requirements.txt
telethon
python-dotenv
tqdm
pandas
sqlalchemy
psycopg2-binary
dbt-postgres
ultralytics
